refactor: 공컴 폴더에서 적출 (FE-128)
